// src/model/LocationSearch.ts
export class LocationSearch {
    Location_Id: number;
    Location_Type: string;
    Location_Name: string;
    Location_Distance: string;
    Distance_Sort: number;
    Distance: number;
    Google_Location: string;
    Image_Url: string;
    Avatar_Url: string;
    Book_Id: number;
    Total_Bookmark: number;
    Location_Address: string;
    Avg_Price: number;
    Open_Time: string;
    Close_Time: string;
}